♻️ Move scrapers to module
